mocking/src/main/java/sales/SalesApp.java
package sales;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Date;
import java.util.List;

public class SalesApp {

	public void generateSalesActivityReport(String salesId, int maxRow, boolean isNatTrade, boolean isSupervisor) {
		
		SalesDao salesDao = new SalesDao();
		SalesReportDao salesReportDao = new SalesReportDao();
		List<String> headers = null;
		
		List<SalesReportData> filteredReportDataList = new ArrayList<SalesReportData>();
		
		if (salesId == null) {
			return;
		}
		
		Sales sales = salesDao.getSalesBySalesId(salesId);

		if (!sales.isActive()){
			return;
		}
		
		List<SalesReportData> reportDataList = salesReportDao.getReportData(sales);

		if (isNatTrade) {
			headers = Arrays.asList("Sales ID", "Sales Name", "Activity", "Time");
		} else {
			headers = Arrays.asList("Sales ID", "Sales Name", "Activity", "Local Time");
		}
		
		SalesActivityReport report = this.generateReport(headers, reportDataList);
	  if(report != null){
			EcmService.uploadDocument(report.toXml());
		}

		
	}

	private SalesActivityReport generateReport(List<String> headers, List<SalesReportData> reportDataList) {

		return new SalesActivityReport();

	}

}
